+++ src/main/java/shop/tryit/domain/item/ImageService.java
@@ -10,7 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.web.item.ItemFormDto;
+import shop.tryit.domain.item.dto.ItemFormDto;
 
 @Slf4j
 @Service
